Make Socket.IO type-safe
