+++ frontend/src/PollPage.tsx
@@ -2,51 +2,12 @@ import { useQuery } from 'react-query';
 import { pollsApi } from './api';
 import { Button, Stack, Typography } from '@mui/material';
 import { useEffect, useState } from 'react';
-import { Socket, io } from 'socket.io-client';
 import { Poll } from './client';
+import useVoteClient, { VoteClient } from './hooks/useVoteClient';
 
 type PollPageProps = {
   params: { code: string };
 };
-
-interface UpdatePayload {
-  questions: [{ id: number; votes: number }];
-}
-
-interface ServerToClientEvents {
-  update: (payload: UpdatePayload) => void;
-}
-
-interface VotePayload {
-  pollCode: number;
-  questionID: number;
-  answerID: number;
-}
-
-interface ClientToServerEvents {
-  vote: (payload: VotePayload) => void;
-}
-
-type VoteClient = Socket<ClientToServerEvents, ServerToClientEvents>;
-
-function useVoteClient(pollCode: string): VoteClient | null {
-  const [voteClient, setVoteClient] = useState<VoteClient | null>(null);
-  useEffect(() => {
-    const client = io(
-      import.meta.env.PROD
-        ? 'https://app.solidpolls.de/api'
-        : 'http://localhost:3000',
-      { transports: ['websocket'] },
-    );
-    setVoteClient(client);
-    return () => {
-      if (client !== null) {
-        client.close();
-      }
-    };
-  }, [pollCode]);
-  return voteClient;
-}
 
 interface QuestionResultsProps {
   poll: Poll;
@@ -74,7 +35,11 @@ interface QuestionVoterProps {
 
 function QuestionVoter(props: QuestionVoterProps) {
   const onVote = () => {
-    props.voteClient.emit('vote', {});
+    props.voteClient.emit('vote', {
+      pollCode: props.poll.code,
+      questionID: 0,
+      answerID: 0,
+    });
     props.onAfterVote();
   };
   return <Button onClick={onVote}>Vote</Button>;
@@ -86,7 +51,7 @@ export default function PollPage(props: PollPageProps) {
     () => pollsApi.pollControllerFindByCode({ code: props.params.code }),
   );
   const voteClient = useVoteClient(props.params.code);
-  const [questionIndex, setQuestionIndex] = useState(0);
+  const [questionIndex] = useState(0);
   const [voted, setVoted] = useState<Set<number>>(new Set());
   const [numVotes, setNumVotes] = useState(0);
   useEffect(() => {
